Remove hard dependency on watchdog
This is necessary because watchdog does not appear to be maintained
anymore, requires a compiler, and has issues installing on the MacOS
High Sierra.

Watchdog has been moved to an optional dependency and is only used if
Chalice is installed with the watchdog dependency extras modifier like
so: `chalice[watchdog]`.

A layer of abstraction was introduced between the watchdog observer
class and the reloader which hides the implementation of the file
observer. One implementation uses the watchdog observer class, which is
used if `import watchdog` succeeds. The fallback implementation uses the
stat system call to check the mtime of each file in the project
directory every second. If the mtime changes then the server is
refreshed.